dinner-party/get_parties/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub
git+https://github.com/anjmittu/dinner-party-database
